+++ libs/components/src/lib/components/overlay/components/bottom-sheet/components/bottom-sheet-container/bottom-sheet-container.component.ts
@@ -40,14 +40,14 @@ import { BottomSheetConfig } from '../../types';
 })
 export class BottomSheetContainerComponent extends BottomSheetContainerBaseComponent {
   constructor(
-    elementRef: ElementRef,
+    elementRef: ElementRef<HTMLElement>,
     focusTrapFactory: FocusTrapFactory,
     @Optional() @Inject(DOCUMENT) document: Document,
     @Inject(BOTTOM_SHEET_CONFIG)
     bottomSheetConfig: BottomSheetConfig,
     checker: InteractivityChecker,
     ngZone: NgZone,
-    overlayRef: OverlayRef,
+    public override overlayRef: OverlayRef,
     focusMonitor?: FocusMonitor,
   ) {
     super(elementRef, focusTrapFactory, document, bottomSheetConfig, checker, ngZone, overlayRef, focusMonitor);