+++ libs/components/src/lib/components/overlay/components/bottom-sheet/partials/bottom-sheet-container-base/bottom-sheet-container-base.component.ts
@@ -12,14 +12,14 @@ export abstract class BottomSheetContainerBaseComponent extends CdkDialogContain
   readonly _animatedLifecycle = inject(ANIMATED_LIFECYCLE_TOKEN);
 
   constructor(
-    public elementRef: ElementRef,
+    public elementRef: ElementRef<HTMLElement>,
     focusTrapFactory: FocusTrapFactory,
     @Optional() @Inject(DOCUMENT) _document: Document,
     @Inject(BOTTOM_SHEET_CONFIG)
     bottomSheetConfig: BottomSheetConfig,
     interactivityChecker: InteractivityChecker,
     ngZone: NgZone,
-    overlayRef: OverlayRef,
+    public overlayRef: OverlayRef,
     focusMonitor?: FocusMonitor,
   ) {
     super(