+++ libs/components/src/lib/components/overlay/components/dialog/partials/dialog-container-base/dialog-container-base.component.ts
@@ -12,14 +12,14 @@ export abstract class DialogContainerBaseComponent extends CdkDialogContainer<Di
   readonly _animatedLifecycle = inject(ANIMATED_LIFECYCLE_TOKEN);
 
   constructor(
-    public elementRef: ElementRef,
+    public elementRef: ElementRef<HTMLElement>,
     focusTrapFactory: FocusTrapFactory,
     @Optional() @Inject(DOCUMENT) _document: Document,
     @Inject(DIALOG_CONFIG)
     dialogConfig: DialogConfig,
     interactivityChecker: InteractivityChecker,
     ngZone: NgZone,
-    overlayRef: OverlayRef,
+    public overlayRef: OverlayRef,
     focusMonitor?: FocusMonitor,
   ) {
     super(