+++ libs/query/src/lib/query-store/query-store.ts
@@ -1,3 +1,4 @@
+import { fromEvent } from 'rxjs';
 import { AnyQuery } from '../query';
 
 export class QueryStore {
@@ -8,8 +9,12 @@ export class QueryStore {
     private _config?: {
       enableChangeLogging?: boolean;
       enableGarbageCollectorLogging?: boolean;
+      autoRefreshQueriesOnWindowFocus?: boolean;
+      enableSmartPolling?: boolean;
     },
-  ) {}
+  ) {
+    this._initSmartQueryHandling();
+  }
 
   add(id: string, query: AnyQuery) {
     this._store.set(id, query);
@@ -27,6 +32,68 @@ export class QueryStore {
     this._store.delete(id);
 
     this._logState(id, null, 'REMOVE');
+  }
+
+  forEach(callback: (value: AnyQuery, key: string) => void) {
+    for (const [key, query] of this._store) {
+      callback(query, key);
+    }
+  }
+
+  refreshQueriesInUse(config?: { ignoreCacheValidity?: boolean; purgeUnused?: boolean }) {
+    const { ignoreCacheValidity, purgeUnused } = config ?? {};
+
+    for (const [key, query] of this._store) {
+      if (
+        query.isInUse &&
+        (query.isExpired || ignoreCacheValidity) &&
+        query.autoRefreshOnConfig.queryClientDefaultHeadersChange
+      ) {
+        query.execute({ skipCache: true, _triggeredVia: 'auto' });
+      } else if (purgeUnused && !query.isInUse) {
+        this.remove(key);
+      }
+    }
+  }
+
+  private _initSmartQueryHandling() {
+    const windowBlur$ = fromEvent<Event>(window, 'blur');
+    const windowFocus$ = fromEvent<Event>(window, 'focus');
+
+    windowBlur$.subscribe(() => {
+      this._stopGarbageCollector();
+
+      if (this._config?.enableSmartPolling) {
+        this.forEach((query) => {
+          if (!query.isPolling || !query._enableSmartPolling) {
+            return;
+          }
+
+          query.pausePolling();
+        });
+      }
+    });
+
+    windowFocus$.subscribe(() => {
+      if (this._config?.enableSmartPolling || this._config?.autoRefreshQueriesOnWindowFocus) {
+        this.forEach((query) => {
+          if (this._config?.enableSmartPolling && query._isPollingPaused) {
+            query.resumePolling();
+          }
+
+          if (
+            this._config?.autoRefreshQueriesOnWindowFocus &&
+            query.isExpired &&
+            query.isInUse &&
+            query.autoRefreshOnConfig.windowFocus
+          ) {
+            query.execute({ skipCache: true, _triggeredVia: 'auto' });
+          }
+        });
+      }
+
+      this._initGarbageCollector();
+    });
   }
 
   private _logState(key: string | null, item: AnyQuery | null, operation: string) {
@@ -67,7 +134,7 @@ export class QueryStore {
     this._logGarbageCollector('Collecting...');
 
     this._store.forEach((item, key) => {
-      if (item.isExpired) {
+      if (item.isExpired && !item.isInUse) {
         this.remove(key);
       }
     });