+++ libs/types/src/lib/api/Identity/index.ts
@@ -1,1 +1,0 @@
-export { IdentityView } from './identity.view';