+++ libs/types/src/lib/api/Statistic/statisticValue.view.ts
@@ -1,8 +1,0 @@
-// Generated Interface, do not change manually!
-
-export interface StatisticValueView {
-    value: number | null;
-    totalMatchGames: number;
-}
-
-export default StatisticValueView;